+++ src/app/core/components/index.ts
@@ -1,5 +1,3 @@
-import InputComponent from './input.component';
-import FormAuthComponent from './form-auth.component';
 import HeaderComponent from './header.component';
 import CardComponent from './card.component';
 import FooterComponent from './footer.component';
@@ -9,8 +7,6 @@ import CardBrandCarouselComponent from './card-brand-carousel.component';
 
 export {
   HeaderComponent,
-  FormAuthComponent,
-  InputComponent,
   CardComponent,
   FooterComponent,
   BannerComponent,
