+++ src/app/core/models/entities/index.ts
@@ -1,0 +1,2 @@
+import { User } from './user.entity';
+export { User };
