+++ src/app/core/models/entities/user.entity.ts
