+++ src/app/features/auth/components/index.ts
@@ -1,0 +1,4 @@
+import FormAuthComponent from './form-auth.component';
+import InputComponent from './input.component';
+
+export { FormAuthComponent, InputComponent };
