+++ tsconfig.json
@@ -23,10 +23,15 @@
     "skipLibCheck": true,
     "paths": {
       "@app/core/components": ["./src/app/core/components/index.ts"],
+      "@app/core/interceptors": ["./src/app/core/interceptors/index.ts"],
+      "@app/core/services": ["./src/app/core/services/index.ts"],
+      "@app/core/models/entities": ["./src/app/core/models/entities/index.ts"],
+      "@app/features/auth/components": [
+        "./src/app/features/auth/components/index.ts"
+      ],
       "@app/services": ["./src/app/services/index.ts"],
       "@app/shared/utils": ["src/app/shared/utils/index.ts"],
-      "@app/shared/interfaces": ["src/app/shared/interfaces/index.ts"],
-      "@app/shared/interceptors": ["src/app/shared/interceptors/index.ts"]
+      "@app/shared/interfaces": ["src/app/shared/interfaces/index.ts"]
     }
   },
   "angularCompilerOptions": {
